% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_nba.R
\name{cor_nba}
\alias{cor_nba}
\title{A  Correlation martix Function}
\usage{
cor_nba(dataset, y)
}
\arguments{
\item{dataset}{is the dataset of NBA}

\item{y}{is the Year that we are going to figure it out}
}
\description{
This function allows you to compute the correlation matrix of numeric variables in the datasets
}
\examples{
cor_nba()
}
\keyword{,matrix}
\keyword{Correlatiin}
