% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest.R
\name{oldest}
\alias{oldest}
\title{A oldest Function}
\usage{
oldest(dataset, y)
}
\arguments{
\item{dataset}{is the dataset of NBA}

\item{y}{is the Year that we are going to figure it out}
}
\description{
This function allows you to find the oldest player of that year
}
\examples{
oldest()
}
\keyword{age}
