% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSM.R
\name{TSMax}
\alias{TSMax}
\title{A highest TS Function}
\usage{
TSMax(dataset, y)
}
\arguments{
\item{dataset}{is NBA datasets}

\item{y}{is the Year that we are going to figure it out}
}
\description{
This function allows you to find the player who has the highest TS of that year
}
\examples{
TSMax(data, 2000)
}
\keyword{TS}
\keyword{highest}
\keyword{points}
